+++ src/main/java/gr/aueb/android/barista/server/dto/BatteryLevelDTO.java
@@ -7,12 +7,12 @@
  * Role:
  * Description:
  */
-package gr.aueb.android.barista.rest.dto;
+package gr.aueb.android.barista.server.dto;
 
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import gr.aueb.android.barista.core.model.BatteryLevel;
-import gr.aueb.android.barista.rest.mapper.CommandMapper;
+import gr.aueb.android.barista.server.mapper.CommandMapper;
 
 @JsonTypeName("BatteryLevel")
 public class BatteryLevelDTO extends CommandDTO<BatteryLevel>{
