+++ src/main/java/gr/aueb/android/barista/server/dto/CommandDTO.java
@@ -1,8 +1,7 @@
-package gr.aueb.android.barista.rest.dto;
+package gr.aueb.android.barista.server.dto;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import com.fasterxml.jackson.annotation.JsonTypeName;
 import gr.aueb.android.barista.core.model.*;
 
 /**
