+++ src/main/java/gr/aueb/android/barista/server/dto/PmGrantDTO.java
@@ -7,11 +7,11 @@
  * Role:
  * Description:
  */
-package gr.aueb.android.barista.rest.dto;
+package gr.aueb.android.barista.server.dto;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import gr.aueb.android.barista.core.model.PmGrant;
-import gr.aueb.android.barista.rest.mapper.CommandMapper;
+import gr.aueb.android.barista.server.mapper.CommandMapper;
 
 @JsonTypeName("PmGrant")
 public class PmGrantDTO extends CommandDTO<PmGrant> {
