+++ src/main/java/gr/aueb/android/barista/server/dto/PmRevokeDTO.java
@@ -7,11 +7,11 @@
  * Role:
  * Description:
  */
-package gr.aueb.android.barista.rest.dto;
+package gr.aueb.android.barista.server.dto;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import gr.aueb.android.barista.core.model.PmRevoke;
-import gr.aueb.android.barista.rest.mapper.CommandMapper;
+import gr.aueb.android.barista.server.mapper.CommandMapper;
 
 @JsonTypeName("PmRevoke")
 public class PmRevokeDTO extends CommandDTO<PmRevoke>{
