+++ src/test/java/gr/aueb/android/barista/server/impl/CommandResourceImplTest.java
@@ -1,13 +1,13 @@
-package gr.aueb.android.barista.rest.impl;
+package gr.aueb.android.barista.server.impl;
 
 import gr.aueb.android.barista.core.executor.CommandExecutorFactory;
 import gr.aueb.android.barista.core.executor.CommandExecutorImplStub;
 import gr.aueb.android.barista.core.model.DimensionUnit;
 import gr.aueb.android.barista.core.emulator.EmulatorManager;
-import gr.aueb.android.barista.rest.dto.CommandDTO;
-import gr.aueb.android.barista.rest.dto.WmDensityDTO;
-import gr.aueb.android.barista.rest.dto.WmSizeDTO;
-import gr.aueb.android.barista.rest.dto.WmSizeResetDTO;
+import gr.aueb.android.barista.server.dto.CommandDTO;
+import gr.aueb.android.barista.server.dto.WmDensityDTO;
+import gr.aueb.android.barista.server.dto.WmSizeDTO;
+import gr.aueb.android.barista.server.dto.WmSizeResetDTO;
 import gr.aueb.android.barista.server.BaristaApplication;
 import gr.aueb.android.barista.server.MyObjectMapperProvider;
 import org.glassfish.jersey.client.ClientConfig;
